Move the source file to `src/index.ts`.
