+++ packages/leb128/deno.json
@@ -1,7 +1,7 @@
 {
   "name": "@commonfabric/leb128",
   "exports": {
-    ".": "./mod.ts"
+    ".": "./src/index.ts"
   },
   "tasks": {
     "test": "deno test"
